[Session] Moving Session classes to a new Manager namespace for better organization.
