+++ src/Valkyrja/Application/Env.php
@@ -60,8 +60,8 @@ use Valkyrja\Mail\Contract\Mailer;
 use Valkyrja\Mail\MailgunMailer;
 use Valkyrja\Orm\Contract\Manager;
 use Valkyrja\Orm\MysqlManager;
-use Valkyrja\Session\Contract\Session;
-use Valkyrja\Session\PhpSession;
+use Valkyrja\Session\Manager\Contract\Session;
+use Valkyrja\Session\Manager\PhpSession;
 use Valkyrja\Sms\Messenger\Contract\Messenger;
 use Valkyrja\Sms\Messenger\VonageMessenger;
 use Valkyrja\View\Renderer\Contract\Renderer;
