+++ src/Valkyrja/Auth/Provider/ServiceProvider.php
@@ -34,7 +34,7 @@ use Valkyrja\Crypt\Contract\Crypt;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Jwt\Contract\Jwt;
 use Valkyrja\Orm\Contract\Manager;
-use Valkyrja\Session\Contract\Session;
+use Valkyrja\Session\Manager\Contract\Session;
 
 /**
  * Class ServiceProvider.
