+++ src/Valkyrja/Session/Manager/Contract/Session.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Session\Contract;
+namespace Valkyrja\Session\Manager\Contract;
 
 use Valkyrja\Session\Exception\InvalidCsrfToken;
