+++ src/Valkyrja/Session/Manager/CookieSession.php
@@ -11,30 +11,31 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Session;
+namespace Valkyrja\Session\Manager;
 
-use JsonException;
 use Override;
-use Valkyrja\Cache\Contract\Cache;
+use Valkyrja\Crypt\Contract\Crypt;
+use Valkyrja\Crypt\Exception\CryptException;
+use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Session\Data\CookieParams;
 use Valkyrja\Session\Exception\SessionStartFailure;
-use Valkyrja\Type\BuiltIn\Support\Arr;
 
 use function headers_sent;
 use function session_start;
 
 /**
- * Class CacheSession.
+ * Class CookieSession.
  *
  * @author Melech Mizrachi
  */
-class CacheSession extends PhpSession
+class CookieSession extends PhpSession
 {
     /**
-     * CacheSession constructor.
+     * CookieSession constructor.
      */
     public function __construct(
-        protected Cache $cache,
+        protected Crypt $crypt,
+        protected ServerRequest $request,
         CookieParams $cookieParams,
         string|null $sessionId = null,
         string|null $sessionName = null
@@ -49,7 +50,7 @@ class CacheSession extends PhpSession
     /**
      * @inheritDoc
      *
-     * @throws JsonException
+     * @throws CryptException
      */
     #[Override]
     public function start(): void
@@ -61,37 +62,37 @@ class CacheSession extends PhpSession
         }
 
         // If the session failed to start
-        if (
-            ! session_start()
-            || ($cachedData = $this->cache->get($this->getCacheSessionId())) === null
-            || $cachedData === ''
-        ) {
+        if (! session_start()) {
             // Throw a new exception
             throw new SessionStartFailure('The session failed to start!');
         }
 
+        $dataString = $this->request->getCookieParam($this->getId());
+
         // Set the data
         /** @psalm-suppress MixedPropertyTypeCoercion */
-        $this->data = Arr::fromString($cachedData);
+        $this->data = $dataString !== null && $dataString !== ''
+            ? $this->crypt->decryptArray($dataString)
+            : [];
     }
 
     /**
      * @inheritDoc
      *
-     * @throws JsonException
+     * @throws CryptException
      */
     #[Override]
     public function set(string $id, $value): void
     {
         $this->data[$id] = $value;
 
-        $this->updateCacheSession();
+        $this->updateCookieSession();
     }
 
     /**
      * @inheritDoc
      *
-     * @throws JsonException
+     * @throws CryptException
      */
     #[Override]
     public function remove(string $id): bool
@@ -102,7 +103,7 @@ class CacheSession extends PhpSession
 
         unset($this->data[$id]);
 
-        $this->updateCacheSession();
+        $this->updateCookieSession();
 
         return true;
     }
@@ -110,48 +111,46 @@ class CacheSession extends PhpSession
     /**
      * @inheritDoc
      *
-     * @throws JsonException
+     * @throws CryptException
      */
     #[Override]
     public function clear(): void
     {
         parent::clear();
 
-        $this->updateCacheSession();
+        $this->updateCookieSession();
     }
 
     /**
      * @inheritDoc
      *
-     * @throws JsonException
+     * @throws CryptException
      */
     #[Override]
     public function destroy(): void
     {
         parent::destroy();
 
-        $this->updateCacheSession();
-    }
-
-    /**
-     * Get the cache session id.
-     *
-     * @return string
-     */
-    protected function getCacheSessionId(): string
-    {
-        return $this->getId() . '_session';
+        $this->updateCookieSession();
     }
 
     /**
      * Update the cache session.
      *
-     * @throws JsonException
+     * @throws CryptException
      *
      * @return void
      */
-    protected function updateCacheSession(): void
+    protected function updateCookieSession(): void
     {
-        $this->cache->forever($this->getCacheSessionId(), Arr::toString($this->data));
+        setcookie(
+            $this->getId(),
+            $this->crypt->encryptArray($this->data),
+            0,
+            '/',
+            '',
+            false,
+            true
+        );
     }
 }
