+++ src/Valkyrja/Session/Manager/PhpSession.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Session;
+namespace Valkyrja\Session\Manager;
 
 use Override;
 use Valkyrja\Session\Data\CookieParams;
