+++ tests/Unit/Session/Provider/ServiceProviderTest.php
@@ -19,13 +19,13 @@ use Valkyrja\Cache\Contract\Cache;
 use Valkyrja\Crypt\Contract\Crypt;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Log\Contract\Logger;
-use Valkyrja\Session\CacheSession;
-use Valkyrja\Session\Contract\Session as Contract;
-use Valkyrja\Session\CookieSession;
 use Valkyrja\Session\Data\CookieParams;
-use Valkyrja\Session\LogSession;
-use Valkyrja\Session\NullSession;
-use Valkyrja\Session\PhpSession;
+use Valkyrja\Session\Manager\CacheSession;
+use Valkyrja\Session\Manager\Contract\Session as Contract;
+use Valkyrja\Session\Manager\CookieSession;
+use Valkyrja\Session\Manager\LogSession;
+use Valkyrja\Session\Manager\NullSession;
+use Valkyrja\Session\Manager\PhpSession;
 use Valkyrja\Session\Provider\ServiceProvider;
 use Valkyrja\Tests\Unit\Container\Provider\ServiceProviderTestCase;
